requirements.txt
selenium
chromedriver-autoinstaller
python-telegram-bot
python-dotenv
beautifulsoup4
requests





